requirements.txt
pyyaml==6.0.3
duckdb==1.4.0
pydantic==2.11.9
tabulate==0.9.0
pyodbc==5.2.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
